FIX #164: Fort Order Residual Post Cycle
- Purge tables on cycle and guard against attribute error.

+++ cogdb/query.py
@@ -406,7 +406,7 @@ def fort_order_remove_finished(session):
     Deletions will be comitted.
     """
     for fort_order in session.query(FortOrder).order_by(FortOrder.order):
-        if fort_order.system.is_fortified or fort_order.system.is_deferred:
+        if fort_order.system and (fort_order.system.is_fortified or fort_order.system.is_deferred):
             session.delete(fort_order)
 
     session.commit()
@@ -1184,6 +1184,7 @@ def post_cycle_db_cleanup(session):
     Args:
         session: Session on to the db.
     """
+    session.query(FortOrder).delete()
     session.query(Vote).delete()
     session.commit()
